feat: operator fee scalar fix

+++ packages/contracts-bedrock/interfaces/L2/IL1Block.sol
@@ -16,7 +16,7 @@ interface IL1Block {
     function l1FeeOverhead() external view returns (uint256);
     function l1FeeScalar() external view returns (uint256);
     function number() external view returns (uint64);
-    function operatorFeeScalar() external view returns (uint32);
+    function operatorFeeScalar() external view returns (uint64);
     function operatorFeeConstant() external view returns (uint64);
     function sequenceNumber() external view returns (uint64);
     function setL1BlockValues(
